adapt scripts to docker environment and update README.MD

+++ scripts/stop-services.sh
@@ -1,28 +1,38 @@
 #!/bin/bash
+set -e
 
 # Ensure the script is executed from the project root directory
 cd "$(dirname "$0")/.."
 
-
 # Source service paths
 config_path="scripts/service-paths-config.sh"
 if [ -f "$config_path" ]; then
-    source "$config_path"
+  source "$config_path"
 else
-    echo "Configuration file not found at $config_path"
-    exit 1
+  echo "Configuration file not found at $config_path"
+  exit 1
 fi
 
 # Function to remove the wildcard from the path
 remove_wildcard() {
-    echo "${1%*.jar}" # This removes "*.jar" from the end of the string
+  echo "${1%*.jar}" # This removes "*.jar" from the end of the string
 }
 
 # Stop the services based on the JAR path without wildcards
 echo "Stopping services..."
-pkill -f "$(remove_wildcard "$product_comp_service")"
-pkill -f "$(remove_wildcard "$product_service")"
-pkill -f "$(remove_wildcard "$recommendation_service")"
-pkill -f "$(remove_wildcard "$review_service")"
+
+# Check the exit code of docker-compose down
+docker_compose_down_result=$(docker compose down --remove-orphans)
+docker_compose_down_exit_code=$?
+
+if [ $docker_compose_down_exit_code -eq 0 ]; then
+  # Use pkill only if docker-compose down might have left behind processes
+  pkill -f "$(remove_wildcard "$product_comp_service")"
+  pkill -f "$(remove_wildcard "$product_service")"
+  pkill -f "$(remove_wildcard "$recommendation_service")"
+  pkill -f "$(remove_wildcard "$review_service")"
+fi
+
+
 
 echo "All services stopped."
